+++ src/Controller/UsersController.php
@@ -35,15 +35,13 @@ class UsersController extends AbstractController
      */
     public function listAction(): Response
     {
-        // TODO: restrict access to this action to admins
+        // TODO: restrict access to admins
         /*throw $this->createNotFoundException('Permission denied: You have to be logged.');*/
 
         $em = $this->getDoctrine()->getManager();
         $usersRepository = $em->getRepository('App\Entity\Users');
         $users = $usersRepository->findAll();
-
         //dump($users);
-
         return $this->render('users/users_list.html.twig', [
             'controller_name' => 'UsersController',
             'users' => $users,
@@ -55,25 +53,28 @@ class UsersController extends AbstractController
      */
     public function addTotoAction(): Response
     {
-        /* raise an exception if toto is already created -> login is unique */
-        $user = new Users(); // l'utilisateur est encore indépendant de Doctrine
-        $user->setLogin('toto')
-            ->setEncPwd('otot')
-            ->setName('Twopak')
-            ->setSurname('Tom')
-            ->setBirthDate(null)
-            ->setIsAdmin(false);// valeur par défaut mais on force
-        //dump($user);
+        $login = "toto";
         $em = $this->getDoctrine()->getManager();
+        $usersRepository = $em->getRepository('App\Entity\Users');
+        $otherToto = $usersRepository->findBy(array("login" => $login));
 
-        $em->persist($user); // Doctrine devient responsable de l'user
-        $em->flush(); // injection physique dans la BD
-        $this->addFlash('info', "toto has been added");
+        if ($otherToto != null){
+            throw $this->createNotFoundException("$login is already in DB");
+        }else{
+            $user = new Users(); // l'utilisateur est encore indépendant de Doctrine
+            $user->setLogin($login)
+                ->setEncPwd('otot')
+                ->setName('Twopak')
+                ->setSurname('Tom')
+                ->setBirthDate(null)
+                ->setIsAdmin(false);
 
-        //dump($user);
+            $em->persist($user); // Doctrine devient responsable de l'user
+            $em->flush(); // injection physique dans la BD
+            $this->addFlash('info', "$login has been added");
 
-        // on redirige vers une autre action
-        return $this->redirectToRoute('users_list');
+            return $this->redirectToRoute("users_index");
+        }
     }
 
     /**
@@ -81,24 +82,31 @@ class UsersController extends AbstractController
      */
     public function addAdminAction(): Response
     {
-        /* raise an exception if admin is already created -> login is unique */
-        $admin = new Users(); // l'utilisateur est encore indépendant de Doctrine
-        $admin->setLogin('admin')
-            ->setEncPwd('nimda')
-            ->setName('Pujadas')
-            ->setSurname('David')
-            ->setBirthDate(null)
-            ->setIsAdmin(true);
-        //dump($admin);
-
+        $login = "admin";
         $em = $this->getDoctrine()->getManager();
+        $usersRepository = $em->getRepository('App\Entity\Users');
+        $otherAdmin = $usersRepository->findBy(array("login" => $login));
 
-        $em->persist($admin); // Doctrine devient responsable de l'user
-        $em->flush(); // injection physique dans la BD
-        $this->addFlash('info', "admin has been added");
+        if ($otherAdmin != null){
+            throw $this->createNotFoundException("$login is already in DB");
+        }else {
+            $admin = new Users(); // l'utilisateur est encore indépendant de Doctrine
+            $admin->setLogin($login)
+                ->setEncPwd('nimda')
+                ->setName('Pujadas')
+                ->setSurname('David')
+                ->setBirthDate(null)
+                ->setIsAdmin(true);
+            //dump($admin);
 
-        // on redirige vers une autre action
-        return $this->redirectToRoute('users_list');
+            $em = $this->getDoctrine()->getManager();
+            $em->persist($admin); // Doctrine devient responsable de l'user
+            $em->flush(); // injection physique dans la BD
+            $this->addFlash('info', "$login has been added");
+
+            // on redirige vers une autre action
+            return $this->redirectToRoute("users_index");
+        }
     }
 
     /**
@@ -109,21 +117,20 @@ class UsersController extends AbstractController
         $form = $this->createForm(AddUserType::class);
         $form->add('send', SubmitType::class, ['label' => 'Add']);
         // We create the form, add a submit button.
-        // dump($request);
-        $userToAdd = $form->handleRequest($request)->getData();
-        // dump($form);
-        dump($userToAdd);
+        $userToAdd = $form->handleRequest($request)->getData(); // doit être fait avant la conditionnelle
 
         // if we already have a posted form, we treat this one.
         if ($form->isSubmitted()) {
             if ($form->isValid()) {
-                $em = $this->getDoctrine()->getManager();
                 $userLogin = $userToAdd->getLogin();
+                $em = $this->getDoctrine()->getManager();
                 $em->persist($userToAdd);
                 $em->flush();
                 $this->addFlash('info', "$userLogin as been added");
+            }else {
+                $this->addFlash('info', 'User hasn\'t been added');
+                throw $this->createNotFoundException('Invalid form.');
             }
-            else {$this->addFlash('info', 'User hasn\'t been added');}
             return $this->redirectToRoute("users_index");
         }
         else {
@@ -141,42 +148,46 @@ class UsersController extends AbstractController
     public function chooseAction(Request $request): Response
     {
         $em = $this->getDoctrine()->getManager();
-
         $usersRepository = $em->getRepository('App\Entity\Users');
         $users = $usersRepository->findAll();
-        // si il y a au moins un user
-        // dump($users);
-        $form = $this->createForm(ChooseUserType::class, $users);
-        $form->add('send', SubmitType::class, ['label' => 'Choose']);
-        // We create the form, add a submit button.
 
-        $data = $form->handleRequest($request)->getData();
-        // dump($data);
+        if ($users == null){
+            throw $this->createNotFoundException("please add some users in order to choose");
+        }else{
+            $form = $this->createForm(ChooseUserType::class, $users);
+            $form->add('send', SubmitType::class, ['label' => 'Choose']);
+            // We create the form, add a submit button.
 
-        // if we already have a posted form, we treat this one.
-        if ($form->isSubmitted()) {
-            if ($form->isValid()) {
-                $userID = $data["user"]->getID();
-                $userLogin =  $data["user"]->getLogin();
-                $actionChosen = $data["action"];
-                $this->addFlash('info', "$userLogin has been chosen");
+            $data = $form->handleRequest($request)->getData();
+            // dump($data);
+            // if we already have a posted form, we treat this one.
+            if ($form->isSubmitted()) {
+                if ($form->isValid()) {
+                    $actionChosen = $data["action"];
+                    $userID = $data["user"]->getID();
+                    $userLogin = $data["user"]->getLogin();
+                    $this->addFlash('info', "$userLogin has been chosen");
 
-                switch ($actionChosen){
-                    case "edit": return $this->redirectToRoute("users_edit", ["id" => $userID]);
-                    case "remove": return $this->redirectToRoute("users_remove", ["id" => $userID]);
-                    default: throw $this->createNotFoundException("$actionChosen/$userID");
+                    switch ($actionChosen) {
+                        case "edit":
+                            return $this->redirectToRoute("users_edit", ["id" => $userID]);
+                        case "remove":
+                            return $this->redirectToRoute("users_remove", ["id" => $userID]);
+                        default:
+                            throw $this->createNotFoundException("$actionChosen/$userID");
+                    }
+                } else{
+                    $this->addFlash('info', 'User hasn\'t been chosen');
+                    throw $this->createNotFoundException('Invalid form.');
                 }
-            }
-            else {
-                $this->addFlash('info', 'User hasn\'t been chosen');
-                throw $this->createNotFoundException('Invalid form.');
+            } else{
+                $myform = $form->createView();
+                return $this->render('users/users_form.html.twig', [
+                    'controller_name' => 'usersController',
+                    'myform' => $myform
+                ]);
             }
         }
-        $myform = $form->createView();
-        return $this->render('users/users_form.html.twig', [
-            'controller_name' => 'usersController',
-            'myform' => $myform
-        ]);
     }
 
     /**
@@ -184,6 +195,7 @@ class UsersController extends AbstractController
      */
     public function removeAction($id): Response
     {
+        // TODO: restrict access to admins
         /*throw $this->createNotFoundException('Permission denied: You have to be logged.');*/
 
         // Default is null for id
@@ -192,10 +204,13 @@ class UsersController extends AbstractController
             $em = $this->getDoctrine()->getManager();
             $usersRepository = $em->getRepository('App\Entity\Users');
             $userToRemove =  $usersRepository->find($id);
-            $em->remove($userToRemove);
-            $em->flush();
-            /*throw $this->createNotFoundException("nothing to do with user $id");*/
-            $this->addFlash('info', "User $id has been removed.");
+            if ($userToRemove != null) {
+                // TODO: check if the user has carts
+                $em->remove($userToRemove);
+                $em->flush();
+                $this->addFlash('info', "User $id has been removed.");
+            }
+            else{throw $this->createNotFoundException('This id is not the pk of a user in this DB.');}
         }
         return $this->redirectToRoute("users_index");
     }
@@ -205,6 +220,7 @@ class UsersController extends AbstractController
      */
     public function editAction($id, Request $request): Response
     {
+        // TODO: restrict access to admins
         /*throw $this->createNotFoundException('Permission denied: You have to be logged.');*/
 
         // Default is null for id
@@ -214,6 +230,8 @@ class UsersController extends AbstractController
             $usersRepository = $em->getRepository('App\Entity\Users');
             $userToEdit =  $usersRepository->find($id);
 
+            if ($userToEdit == null) {throw $this->createNotFoundException('This id is not the pk of a user in this DB.');}
+            // login is unique so an exception is gonna be raised by Doctrine if it already exist
             $form = $this->createForm(AddUserType::class, $userToEdit);
             $form->add('send', SubmitType::class, ['label' => 'Edit']);
             // We create the form, add a submit button.
@@ -228,7 +246,10 @@ class UsersController extends AbstractController
                     $em->flush();
                     $this->addFlash('info', "User $id has been edited");
                 }
-                else {$this->addFlash('info', 'User hasn\'t been edited');}
+                else {
+                    $this->addFlash('info', 'User hasn\'t been edited');
+                    throw $this->createNotFoundException('Invalid form.');
+                }
                 return $this->redirectToRoute("users_list");
             }
             else {
@@ -240,5 +261,4 @@ class UsersController extends AbstractController
             }
         }
     }
-
 }
